+++ tests/test_dynamic_array.cpp
@@ -43,7 +43,7 @@ private:
     
     void printArray(DynamicArray<int>& arr, const std::string& label) {
         std::cout << "    " << label << ": [";
-        for (size_t i = 0; i < static_cast<size_t>(arr.GetSize()); i++) {
+        for (size_t i = 0; i < (arr.GetSize()); i++) {
             std::cout << arr.Get(i);
             if (i < arr.GetSize() - 1) std::cout << ", ";
         }
@@ -55,14 +55,14 @@ public:
         std::cout << "\n--- Тесты конструкторов DynamicArray ---" << std::endl;
         
         DynamicArray<int> arr1;
-        assertEqual(static_cast<size_t>(0), arr1.GetSize(), "Конструктор по умолчанию - размер");
+        assertEqual((0), arr1.GetSize(), "Конструктор по умолчанию - размер");
         
         DynamicArray<int> arr2(5);
-        assertEqual(static_cast<size_t>(5), arr2.GetSize(), "Конструктор с размером - размер");
+        assertEqual((5), arr2.GetSize(), "Конструктор с размером - размер");
         
         int items[] = {1, 2, 3, 4, 5};
         DynamicArray<int> arr3(items, 5);
-        assertEqual(static_cast<size_t>(5), arr3.GetSize(), "Конструктор из массива - размер");
+        assertEqual((5), arr3.GetSize(), "Конструктор из массива - размер");
         assertEqual(1, arr3.Get(0), "Конструктор из массива - элемент 0");
         assertEqual(3, arr3.Get(2), "Конструктор из массива - элемент 2");
         assertEqual(5, arr3.Get(4), "Конструктор из массива - элемент 4");
@@ -71,7 +71,7 @@ public:
         printArray(arr3, "Созданный массив");
         
         DynamicArray<int> arr4(arr3);
-        assertEqual(static_cast<size_t>(5), arr4.GetSize(), "Конструктор копирования - размер");
+        assertEqual((5), arr4.GetSize(), "Конструктор копирования - размер");
         assertEqual(1, arr4.Get(0), "Конструктор копирования - элемент 0");
     }
     
@@ -119,20 +119,20 @@ public:
         std::cout << "    Исходный массив: [1, 2, 3, 4, 5]" << std::endl;
         
         arr.Resize(8);
-        assertEqual(static_cast<size_t>(8), arr.GetSize(), "Увеличение размера до 8 - новый размер");
+        assertEqual((8), arr.GetSize(), "Увеличение размера до 8 - новый размер");
         assertEqual(1, arr.Get(0), "Увеличение размера - элементы сохранены");
         assertEqual(5, arr.Get(4), "Увеличение размера - последний элемент сохранен");
         assertEqual(0, arr.Get(7), "Увеличение размера - новые элементы обнулены");
         std::cout << "    После Resize(8): [1, 2, 3, 4, 5, 0, 0, 0]" << std::endl;
         
         arr.Resize(3);
-        assertEqual(static_cast<size_t>(3), arr.GetSize(), "Уменьшение размера до 3 - новый размер");
+        assertEqual((3), arr.GetSize(), "Уменьшение размера до 3 - новый размер");
         assertEqual(1, arr.Get(0), "Уменьшение размера - первые элементы сохранены");
         assertEqual(3, arr.Get(2), "Уменьшение размера - обрезано корректно");
         std::cout << "    После Resize(3): [1, 2, 3]" << std::endl;
         
         try {
-            arr.Resize(static_cast<size_t>(-1));
+            arr.Resize((-1));
             assertTrue(false, "Resize с отрицательным размером - должно быть исключение");
         } catch (const IndexOutOfRangeException&) {
             assertTrue(true, "Resize с отрицательным размером - исключение перехвачено");
@@ -143,7 +143,7 @@ public:
         std::cout << "\n--- Тесты граничных случаев DynamicArray ---" << std::endl;
         
         DynamicArray<int> empty;
-        assertEqual(static_cast<size_t>(0), empty.GetSize(), "Пустой массив - размер");
+        assertEqual((0), empty.GetSize(), "Пустой массив - размер");
         std::cout << "    Пустой массив создан успешно" << std::endl;
         
         DynamicArray<int> single(1);
@@ -156,7 +156,7 @@ public:
         assertEqual(7, singleFromArray.Get(0), "Массив из одного элемента из массива");
         
         DynamicArray<int> zeroSize(0);
-        assertEqual(static_cast<size_t>(0), zeroSize.GetSize(), "Массив нулевого размера");
+        assertEqual((0), zeroSize.GetSize(), "Массив нулевого размера");
     }
     
     void testAssignment() {
@@ -170,7 +170,7 @@ public:
         std::cout << "    arr2: []" << std::endl;
         
         arr2 = arr1;
-        assertEqual(static_cast<size_t>(3), arr2.GetSize(), "Присваивание - размер скопирован");
+        assertEqual((3), arr2.GetSize(), "Присваивание - размер скопирован");
         assertEqual(10, arr2.Get(0), "Присваивание - элемент 0 скопирован");
         assertEqual(30, arr2.Get(2), "Присваивание - элемент 2 скопирован");
